IGAPP-114: Fix move of navigation types
